Dev 3.0

+++ src/Adapters/BlinkerPRO.h
@@ -64,7 +64,7 @@ class BlinkerPRO : public BlinkerStream
         int  duerPrint(const String & data, bool report = false);
         // void aliType(const String & type);
         void begin(const char* _deviceType);
-        int autoPrint(uint32_t id);
+        int autoPrint(unsigned long id);
         // bool autoPrint(char *name, char *type, char *data);
         // bool autoPrint(char *name1, char *type1, char *data1, \
         //             char *name2, char *type2, char *data2);
@@ -1028,7 +1028,7 @@ void BlinkerPRO::begin(const char* _type)
     mDNSInit();
 }
 
-int BlinkerPRO::autoPrint(uint32_t id)
+int BlinkerPRO::autoPrint(unsigned long id)
 {
     String payload = BLINKER_F("{\"data\":{\"set\":{");
     payload += BLINKER_F("\"trigged\":true,\"autoData\":{");
