+++ src/Adapters/BlinkerPROAIR202.h
@@ -62,7 +62,7 @@ class BlinkerProAIR202 : public BlinkerStream
         // void duerType(const String & type);
         void begin(const char* _deviceType, String _imei);
         void initStream(Stream& s, bool state, blinker_callback_t func);
-        int autoPrint(uint32_t id);
+        int autoPrint(unsigned long id);
         char * deviceName();
         char * authKey() { return AUTHKEY_GPRS; }
         int init() { return isMQTTinit; }
@@ -833,7 +833,7 @@ void BlinkerProAIR202::initStream(Stream& s, bool state, blinker_callback_t func
     // strcpy(_imei, imei.c_str());
 }
 
-int BlinkerProAIR202::autoPrint(uint32_t id)
+int BlinkerProAIR202::autoPrint(unsigned long id)
 {
     String payload = BLINKER_F("{\"data\":{\"set\":{");
     payload += BLINKER_F("\"trigged\":true,\"autoData\":{");
